qca-cmn-fw: Split PHY HR/QR Bring up Host Changes
In splitphy HR/QR bring up sequence, PHYA0 and PHYA1 are
configured with different rates, leading to a crash of
the AP/STA.
The root cause of the crash is the mismatch in PHY rates
between PHYA0 and PHYA1. When the rates are different,
it causes an inconsistency in the PHY configuration,
leading to a Ucode crash.

Added changes to send the chanbw vdev param to FW.
FW sets this rate to

Change-Id: I13670722565c1eb27bd2090899f9faece5251922